sleepy time
Cleanup still in progress
List of broken shit:
-All slash commands
-All background loops(need testing)
-All housekeeping shit
-All information on bot(not recognizing background loops running in extensions)
-Make more informational logchannel messages (make togglable with config idiot)
-Fix readme(issues on github)
-Make mayor channel show election time (formula is somewhere)
