+++ utils.py
@@ -138,4 +138,3 @@ def mayorgraphing():
     graphurl = f"https://image-charts.com/chart?&chco=CFECF7%2C27c9c2&chd=t:{performance[0]},{performance[1]},{performance[2]},{performance[3]},{performance[4]}&chs=700x300&cht=bvs&chtt=SkyBlock%20Mayor%20Votes&chxl=0%3A|{objectslist[0]}|{objectslist[1]}|{objectslist[2]}|{objectslist[3]}|{objectslist[4]}&chxs=1N*2s*{modifier},000000&chxt=x,y&chg=20,50,5,5,CECECE"
     return graphurl
     
-=======
